Use jwt for storing gatewayAttempts
When using session, secondary calls will overwrite the gatewayAttempts property,
leading to possibly infinite redirect loops. This is especially obvious when loading
Directory pages with lots of images, all requiring a trip to the cas gateway server.
Setting the gatewayAttempts number on the url instead prevents this. The token is
encoded as a jwt.