Disable erasing anonymous-users for gateway logins
If anonymous-user is not set, every request will require a round trip
to the CAS server. This causes problems primarily for directory-web and
files-web, where the gatewayAttempts session property gets overwritten by competing
requests to avatar thumbnail pictures from an organization list.

If a user logs in, the cookie will still be reset, and authenticated content should
become available.